feat: keep track of filter state instead of filtered animals
This allows the app to have different characteristics based on which
filter is being applied currently, e.g. disable the currently active
filter button.
